+++ questionbox/project/forms.py
@@ -1,0 +1,1 @@
+class UserForm(forms.ModelForm)
